pyproject.toml
[project]
name = "airflow"
version = "0.1.0"
requires-python = "==3.12.*"
dependencies = [
    "dbt-core==1.8.9",
    "dbt-duckdb==1.8.2",
    "duckdb==1.1.3",
    "astronomer-cosmos==1.8.2",
    "pandas==2.2.3",
    "openpyxl==3.1.5",
]

[project.optional-dependencies]
dev = [
    "astro-run-dag",
]
